// src/graph/types.ts
export type Input<T> = { input: T };
export type Where<T> = { where: T };

export type Maybe<T> = T | null;
// ...
export interface CreateDictionaryData {
  name: string;
}

export interface BatchDictionaryData {
  id: number;

  name: Maybe<string>;

  values: DictionaryValueData[];
}

export interface DictionaryValueData {
  actionType: ActionType;

  id: Maybe<number>;

  value: Maybe<string>;
}

export interface DeleteDictionaryData {
  id: number;
}

export enum IntentTriggerType {
  Plain = "PLAIN",
  Combination = "COMBINATION",
}

export enum ActionType {
  Create = "CREATE",
  Update = "UPDATE",
  Delete = "DELETE",
}